Refactor oversized shared instruction modules
